stuff for misc function testing

+++ core/moods.h
@@ -3,16 +3,17 @@
 
 
 #include <vector>
-#include <vector>
 #include <cmath>
+#include <cstdint>
 #include <limits>
 
 
 // for data about position in a sequence
+// TODO: replace with size_t 
 typedef unsigned long long position_t;
 
 // for bit parallel magic
-typedef uint_fast32_t bits_t;
+typedef unsigned long bits_t;
 
 // matrix scores
 typedef std::vector<std::vector<double> > score_matrix;
